+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/event/processor/IdentityDeleteProcessor.java
@@ -9,10 +9,11 @@ import eu.bcvsolutions.idm.acc.dto.filter.AccIdentityAccountFilter;
 import eu.bcvsolutions.idm.acc.service.api.AccIdentityAccountService;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
 import eu.bcvsolutions.idm.core.api.event.CoreEvent;
+import eu.bcvsolutions.idm.core.api.event.CoreEventProcessor;
 import eu.bcvsolutions.idm.core.api.event.DefaultEventResult;
 import eu.bcvsolutions.idm.core.api.event.EntityEvent;
 import eu.bcvsolutions.idm.core.api.event.EventResult;
-import eu.bcvsolutions.idm.core.api.event.processor.AbstractIdentityProcessor;
+import eu.bcvsolutions.idm.core.api.event.processor.IdentityProcessor;
 import eu.bcvsolutions.idm.core.model.event.IdentityEvent.IdentityEventType;
 
 /**
@@ -23,7 +24,9 @@ import eu.bcvsolutions.idm.core.model.event.IdentityEvent.IdentityEventType;
  */
 @Component("accIdentityDeleteProcessor")
 @Description("Ensures referential integrity. Cannot be disabled.")
-public class IdentityDeleteProcessor extends AbstractIdentityProcessor {
+public class IdentityDeleteProcessor
+		extends CoreEventProcessor<IdmIdentityDto> 
+		implements IdentityProcessor {
 	
 	public static final String PROCESSOR_NAME = "identity-delete-processor";
 	private final AccIdentityAccountService identityAccountService;