+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/event/EntityEvent.java
@@ -4,6 +4,7 @@ import java.io.Serializable;
 import java.util.Map;
 
 import org.springframework.core.ResolvableTypeProvider;
+import org.springframework.util.Assert;
 
 import eu.bcvsolutions.idm.core.api.dto.BaseDto;
 import eu.bcvsolutions.idm.core.api.entity.BaseEntity;
@@ -95,4 +96,18 @@ public interface EntityEvent<E extends Serializable> extends ResolvableTypeProvi
 	 * @return
 	 */
 	Integer getProcessedOrder();
+	
+	
+	/**
+	 * Returns true, if event's type equals given eventType.
+	 * 
+	 * @param event
+	 * @param eventType
+	 * @return
+	 */
+	default boolean hasType(EventType eventType) {
+		Assert.notNull(eventType);
+		//
+		return eventType.name() == getType().name();
+	}
 }