+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/event/processor/IdentityProcessor.java
@@ -1,0 +1,14 @@
+package eu.bcvsolutions.idm.core.api.event.processor;
+
+import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
+import eu.bcvsolutions.idm.core.api.event.EntityEventProcessor;
+
+/**
+ * Identity processors should implement this interface.
+ * 
+ * @author Radek Tomiška
+ *
+ */
+public interface IdentityProcessor extends EntityEventProcessor<IdmIdentityDto> {
+	
+}